Debian patch 25_duplicate_iftable
Patch from
http://sourceforge.net/tracker/index.php?func=detail&aid=1513191&group_id=12694&atid=312694 :

This patch fixes bug #1238981 for me. If a new device
will be added, the old device are first checked. If
there's an old device with the same name is found, the
old device is removed from the list. This fixes the
problem that i have tons of ppp0 interfaces in the
snmpd ifTable.

Signed-of-by: Sven Schnelle <svens@gmx.de>