snmpd: disable logging of debug messages
Bug 7395
Do not put snmpd debug messages into syslog.
The container code logs a debug message if an attempt is made to put
two entries into same container with same key. This is expected for the
case of two interfaces with same MAC address (examples are VLAN, bridge, or
bonding) that result in same link local IPv6 address.