optimize insertion of binary array container
The binary array is one of the most common container types in net-snmp
but each insert causes a lookup and a resort of the whole array.
Instead, of sorting use the lookup process to determine the location
in the array to insert, and optimize for the most common case.

Most containers insert items in order, so it makes sense to optimize
for the case of appending. This especially important for huge tables
like the route table on a router.
---
 snmplib/container_binary_array.c |   54 ++++++++++++++++++++++++++-----------
 1 files changed, 38 insertions(+), 16 deletions(-)