Support netlink for reading routes
Adds a module that uses netlink to get route information and directly
populates the route cache (inetCidrRouteTable and ipCidrRouteTable).

Also changes the route table cache policy to reset the expire timer on
access so that the cache does not expire in the middle of walking a
large route table.

Fixes BZ8199